backend/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
google-genai==0.1.0
aiohttp==3.7.4.post0
psycopg2==2.9.1
python-dotenv==1.0.0
python-json-logger==2.0.7